feat: rename cli-ux global to ux global
