+++ package.json
@@ -105,14 +105,14 @@
     "./errors": "./lib/errors/index.js",
     "./execute": "./lib/execute.js",
     "./flags": "./lib/flags.js",
-    "./flush": "./lib/cli-ux/flush.js",
+    "./flush": "./lib/ux/flush.js",
     "./handle": "./lib/errors/handle.js",
     "./help": "./lib/help/index.js",
     "./interfaces": "./lib/interfaces/index.js",
     "./performance": "./lib/performance.js",
     "./run": "./lib/main.js",
     "./settings": "./lib/settings.js",
-    "./ux": "./lib/cli-ux/index.js"
+    "./ux": "./lib/ux/index.js"
   },
   "repository": "oclif/core",
   "oclif": {
