+++ src/config/config.ts
@@ -15,7 +15,7 @@ import ModuleLoader from '../module-loader'
 import {getHelpFlagAdditions} from '../help'
 import {Class, Loadable, Cached, Flag, Arg as CommandArg} from '../command'
 import {CompletableOptionFlag, Arg} from '../interfaces/parser'
-import {stdout} from '../cli-ux/stream'
+import {stdout} from '../ux/stream'
 import Performance from '../performance'
 import settings from '../settings'
 import {userInfo as osUserInfo} from 'node:os'
