+++ src/main.ts
@@ -6,7 +6,7 @@ import * as Interfaces from './interfaces'
 import {URL} from 'url'
 import Config from './config'
 import {getHelpFlagAdditions, loadHelpClass, normalizeArgv} from './help'
-import {stdout} from './cli-ux/stream'
+import {stdout} from './ux/stream'
 import Performance from './performance'
 
 const debug = require('debug')('oclif:main')
