+++ src/screen.ts
@@ -1,4 +1,4 @@
-import {stdout, stderr} from './cli-ux/stream'
+import {stdout, stderr} from './ux/stream'
 import settings from './settings'
 
 function termwidth(stream: any): number {
