+++ src/ux/global.d.ts
@@ -1,5 +1,5 @@
 declare namespace NodeJS {
   interface Global {
-    'cli-ux': any;
+    ux: any;
   }
 }
