+++ src/ux/styled/object.ts
