+++ src/ux/styled/table.ts
