+++ test/cli-ux/index.test.ts
@@ -1,4 +1,4 @@
-import ux from '../../src/cli-ux'
+import ux from '../../src/ux'
 
 import {expect, fancy} from './fancy'
 const hyperlinker = require('hyperlinker')
