+++ test/cli-ux/styled/progress.test.ts
@@ -1,5 +1,5 @@
 import {expect, fancy} from 'fancy-test'
-import ux from '../../../src/cli-ux'
+import ux from '../../../src/ux'
 
 describe('progress', () => {
   // single bar
