+++ test/cli-ux/styled/table.e2e.ts
@@ -1,5 +1,5 @@
 import {expect, fancy} from 'fancy-test'
-import ux from '../../../src/cli-ux'
+import ux from '../../../src/ux'
 
 describe('styled/table', () => {
   describe('null/undefined handling', () => {
